# Import Performance optimizations: NativeArrays for buffer access, multithreading changes and asset importing speedups