--- a/Editor/Scripts/GLTFImporter.cs
+++ b/Editor/Scripts/GLTFImporter.cs
@@ -263,7 +263,8 @@ namespace UnityGLTF
             AnimationClip[] animations = null;
             Mesh[] meshes = null;
 
-            var uniqueNames = new List<string>() { "main asset" };
+	        var uniqueNames = new Dictionary<int, int>(100);
+	        uniqueNames.Add("main asset".GetHashCode(), 0);
             EnsureShadersAreLoaded();
 
 #if UNITY_2020_2_OR_NEWER && !UNITY_2021_3_OR_NEWER
@@ -273,9 +274,17 @@ namespace UnityGLTF
 
             string GetUniqueName(string desiredName)
             {
-	            var uniqueName = ObjectNames.GetUniqueName(uniqueNames.ToArray(), desiredName);
-	            if (!uniqueNames.Contains(uniqueName)) uniqueNames.Add(uniqueName);
-	            return uniqueName;
+	            var hash = desiredName.GetHashCode();
+	            if (uniqueNames.ContainsKey(hash))
+	            {
+		            uniqueNames[hash]++;
+		            return $"{desiredName} ({uniqueNames[hash]})";
+	            }
+	            else
+	            {
+		            uniqueNames.Add(hash, 0);
+		            return desiredName;
+	            }
             }
 
 #if UNITY_2017_3_OR_NEWER
@@ -807,8 +816,11 @@ namespace UnityGLTF
 	        // run texture verification and warn about wrong configuration
 	        if (!GLTFImporterHelper.TextureImportSettingsAreCorrect(this))
 		        Debug.LogWarning("Some Textures have incorrect linear/sRGB settings. Use the \"Fix All\" button in the importer to adjust.");
-		        
+
+	        if (context.SceneImporter != null)
+		        context.SceneImporter.Dispose();
 		}
+		        
 
         private const string ColorSpaceDependency = nameof(GLTFImporter) + "_" + nameof(PlayerSettings.colorSpace);
         private const string NormalMapEncodingDependency = nameof(GLTFImporter) + "_normalMapEncoding";