--- a/Runtime/Plugins/GLTFSerialization/GLTFSerialization.asmdef
+++ b/Runtime/Plugins/GLTFSerialization/GLTFSerialization.asmdef
@@ -2,7 +2,10 @@
     "name": "GLTFSerialization",
     "rootNamespace": "",
     "references": [
-        "GUID:84ca1e7ada6c144debb2e417c8dbe9a5"
+        "GUID:84ca1e7ada6c144debb2e417c8dbe9a5",
+        "GUID:e0cd26848372d4e5c891c569017e11f1",
+        "GUID:2665a8d13d1b3f18800f46e256720795",
+        "GUID:d8b63aba1907145bea998dd612889d6b"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
@@ -20,5 +23,5 @@
             "define": "HAVE_MESHOPT_DECOMPRESS"
         }
     ],
-    "noEngineReferences": true
+    "noEngineReferences": false
 }