--- a/Runtime/Scripts/Cache/BufferCacheData.cs
+++ b/Runtime/Scripts/Cache/BufferCacheData.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Unity.Collections;
 
 namespace UnityGLTF.Cache
 {
@@ -7,11 +8,17 @@ namespace UnityGLTF.Cache
 		public uint ChunkOffset { get; set; }
 		public System.IO.Stream Stream { get; set; }
 
+		public NativeArray<byte> bufferData { get; set; }
+
 		public void Dispose()
 		{
 			if (Stream != null)
 			{
+#if !WINDOWS_UWP
+				Stream.Close();
+#else
 				Stream.Dispose();
+#endif				
 				Stream = null;
 			}
 		}