… Runtime/Plugins/GLTFSerialization/Schema/Accessor.cs
… Runtime/Plugins/GLTFSerialization/Schema/SchemaExtensions.cs.meta
… Runtime/Scripts/Cache/MeshCacheData.cs
… Runtime/Scripts/GLTFSceneImporter.cs
… Runtime/Scripts/SceneImporter/ImporterMeshes.cs
… Runtime/Scripts/SceneImporter/ImporterSkinning.cs
… Runtime/Scripts/SceneImporter/ImporterTextures.cs
… Runtime/UnityGLTFScripts.asmdef