- Custom solution for ObjectNames.GetUniqueName when importing assets: Old solution was too slow on gltf's wich a lot of subassets
- Changed buffer access to NativeArrays: When getting bufferData for bufferViews, instead of filling up everytime managed arrays, we now using NativeArrays.GetSubArray from the buffer. Especially on gltf with a large amound of bufferviews, this reducing loading times by a big amount.
- Optimized reading of accessor data and conversion
- Draco decoding: instead of decoding draco meshes in serial, we now start decoding all draco meshes at first and then wating of them do complete before building the nodes and unity meshes.
- Multithreading changes:  when multithreading was enabled, the importing process takes much more time to complete, because of the overhead of C# Tasks. So instead of many very small Task.Run( ...), they are now combined to only a small amount of tasks.  

Some tests:
With the test model from [Issue 486](https://github.com/KhronosGroup/UnityGLTF/issues/486) i was able to reduce the asset importing time on my machine from over 2 minutes to under 4 seconds . In runtime build even to 1.5 seconds.
I also tested a in-house model (30mb) which included draco and ktx:  which was in runtime build with multithreaded 16 seconds down to 3.4 seconds and with multithreaded disabled from 8.3 seconds down to 3.5 seconds.

Smaller gltf files don't get much impact from this changes.